requirements.txt
boto3>=1.34.0
dash>=2.16.0
dash-bootstrap-components>=1.6.0
plotly>=5.20.0
pandas>=2.0.0
requests>=2.31.0
python-dotenv>=1.0.0
wordcloud>=1.9.0
matplotlib>=3.8.0
